fix: add Copy and Clone traits to stucts
- Add `Clone` trait to all structures.
- Add `Copy` trait to most structures. Descriptors won't support
  `Copy` due to `String` types.

Without Copy/Clone traits on the structures, writing supporting
EDID handling functions force the compiler to move ownership of
the data instead of borrowing.

The Clone trait does not change this behavior, but it does allow
for an explicit (and maybe time expensive) secondary set of EDID
data to process.